app/requirements.txt
azure-ai-documentintelligence>=1.0.0b3
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.1
pydantic>=1.10.0
tenacity>=8.2.0
python-dotenv>=1.0.0
